requirements.txt
pystache==0.5.4
PyYAML==3.12
